✨ Split GPUReservations.tsx: extract useGPUDashboardCards, useGPUCalendarState, useGPUReservationForm
Extracts three focused custom hooks from the monolithic GPUReservations component
to reduce the hook count and improve maintainability:

- useGPUDashboardCards: dashboard card state, localStorage persistence, DnD
- useGPUCalendarState: month navigation, calendar-week layout computation
- useGPUReservationForm: reservation form / delete-confirmation state

GPUReservations.tsx is reduced from ~40 hook calls to ~27 by delegating
each logical group to its own hook.

Fixes #21612
