+++ web/src/components/gpu/useGPUDashboardCards.ts
@@ -1,0 +1,85 @@
+import { useState } from 'react'
+import {
+  KeyboardSensor,
+  PointerSensor,
+  useSensor,
+  useSensors,
+  type DragEndEvent,
+} from '@dnd-kit/core'
+import { arrayMove, sortableKeyboardCoordinates } from '@dnd-kit/sortable'
+import { getDefaultCardWidth } from '../cards/cardRegistry'
+import type { GpuDashCard } from './SortableGpuCard'
+import { DEFAULT_GPU_CARDS } from './SortableGpuCard'
+import { safeGetJSON, safeSetJSON } from '../../lib/utils/localStorage'
+
+const GPU_DASHBOARD_STORAGE_KEY = 'gpu-dashboard-tab-cards'
+
+export function useGPUDashboardCards() {
+  const [dashboardCards, setDashboardCards] = useState<GpuDashCard[]>(() => {
+    const stored = safeGetJSON<GpuDashCard[] | string[]>(GPU_DASHBOARD_STORAGE_KEY)
+    if (!stored || stored.length === 0) return DEFAULT_GPU_CARDS
+    // Migrate from old string[] format
+    if (typeof stored[0] === 'string') {
+      const migrated = (stored as string[]).map(type => ({ type, width: getDefaultCardWidth(type) }))
+      safeSetJSON(GPU_DASHBOARD_STORAGE_KEY, migrated)
+      return migrated
+    }
+    return stored as GpuDashCard[]
+  })
+
+  const gpuDashSensors = useSensors(
+    useSensor(PointerSensor, { activationConstraint: { distance: 8 } }),
+    useSensor(KeyboardSensor, { coordinateGetter: sortableKeyboardCoordinates })
+  )
+
+  const dashCardIds = dashboardCards.map((c, i) => `gpu-dash-${c.type}-${i}`)
+
+  const handleAddDashboardCards = (suggestions: Array<{ type: string; title: string; visualization: string; config: Record<string, unknown> }>) => {
+    setDashboardCards(prev => {
+      const updated = [...prev, ...suggestions.map(s => ({ type: s.type, width: getDefaultCardWidth(s.type) }))]
+      safeSetJSON(GPU_DASHBOARD_STORAGE_KEY, updated)
+      return updated
+    })
+  }
+
+  const handleRemoveDashboardCard = (index: number) => {
+    setDashboardCards(prev => {
+      const updated = prev.filter((_, i) => i !== index)
+      safeSetJSON(GPU_DASHBOARD_STORAGE_KEY, updated)
+      return updated
+    })
+  }
+
+  const handleDashCardWidthChange = (index: number, newWidth: number) => {
+    setDashboardCards(prev => {
+      const updated = prev.map((c, i) => i === index ? { ...c, width: newWidth } : c)
+      safeSetJSON(GPU_DASHBOARD_STORAGE_KEY, updated)
+      return updated
+    })
+  }
+
+  const handleDashDragEnd = (event: DragEndEvent) => {
+    const { active, over } = event
+    if (over && active.id !== over.id) {
+      const oldIndex = dashCardIds.indexOf(active.id as string)
+      const newIndex = dashCardIds.indexOf(over.id as string)
+      if (oldIndex !== -1 && newIndex !== -1) {
+        setDashboardCards(prev => {
+          const updated = arrayMove(prev, oldIndex, newIndex)
+          safeSetJSON(GPU_DASHBOARD_STORAGE_KEY, updated)
+          return updated
+        })
+      }
+    }
+  }
+
+  return {
+    dashboardCards,
+    dashCardIds,
+    gpuDashSensors,
+    handleAddDashboardCards,
+    handleRemoveDashboardCard,
+    handleDashCardWidthChange,
+    handleDashDragEnd,
+  }
+}
